requirements.txt
numpy==2.2.5
openpyxl==3.1.5
pandas==2.2.3
pillow==11.2.1
PyMuPDF==1.25.5
